+++ module/model-sql/__index__.ts
@@ -2,9 +2,5 @@ export * from './src/config.ts';
 export * from './src/connection.ts';
 export * from './src/dialect.ts';
 export * from './src/query.ts';
-export * from './src/service-util/bulk.ts';
-export * from './src/service-util/crud.ts';
-export * from './src/service-util/indexed.ts';
-export * from './src/service-util/query.ts';
-export * from './src/service-util/storage.ts';
+export * from './src/service.ts';
 export * from './src/util.ts';
